Change most of the chart's base type to RefCounted

+++ UI/IComboDisplay.cs
@@ -13,7 +13,7 @@ public interface IComboDisplay
     /// <param name="combo">The current combo</param>
     /// <param name="type">The hit type provided</param>
     /// <param name="offset">A Vector2 that offsets the position</param>
-    public void Show(uint combo, HitType type, Vector2? offset);
+    public void Show(long combo, HitType type, Vector2? offset);
 
     /// <summary>
     /// Displays the combo information with the hit type provided, providing access to the anchors as well.
@@ -25,5 +25,5 @@ public interface IComboDisplay
     /// <param name="anchorRight">The right anchor (usually from 0 to 1)</param>
     /// <param name="anchorBottom">The bottom anchor (usually from 0 to 1)</param>
     /// <param name="offset">Where to offset the combo from the anchors, in pixels.</param>
-    public void Show(uint combo, HitType type, float anchorLeft, float anchorTop, float anchorRight, float anchorBottom, Vector2? offset);
+    public void Show(long combo, HitType type, float anchorLeft, float anchorTop, float anchorRight, float anchorBottom, Vector2? offset);
 }
